Upgrade httpparty dependency 0.13.1 -> 0.17.3
`httpparty` version 0.13.1 has a dependency on an
older version of the `json` gem.  This can cause
incompatibility issues on projects that include
gems requiring more recent versions of the `json`
gem.

Beginning with version 0.14.0 of `httpparty`, the
`json` dependency was removed completely.

Upgrading to the latest version of `httpparty`
made sense.